Better handle of winning moves

+++ src/main/java/fr/leblanc/gomoku/repository/GameRepository.java
@@ -5,6 +5,6 @@ import fr.leblanc.gomoku.model.Game;
 import org.springframework.data.repository.CrudRepository;
 
 @Repository
-public interface GameRepository extends CrudRepository<Game, Integer>
+public interface GameRepository extends CrudRepository<Game, Long>
 {
 }
